reformat code by google style code

+++ src/main/java/com/labate/mentoringme/service/quizz/QuizServiceImpl.java
@@ -1,16 +1,5 @@
 package com.labate.mentoringme.service.quizz;
 
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
-import org.modelmapper.ModelMapper;
-import org.springframework.data.domain.Page;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.stereotype.Service;
 import com.labate.mentoringme.dto.QuizResultCheckingDto;
 import com.labate.mentoringme.dto.UserSelectionDto;
 import com.labate.mentoringme.dto.mapper.PageCriteriaPageableMapper;
@@ -32,6 +21,13 @@ import com.labate.mentoringme.repository.QuizRepository;
 import com.labate.mentoringme.repository.QuizResultRepository;
 import com.labate.mentoringme.util.ObjectMapperUtils;
 import lombok.AllArgsConstructor;
+import org.modelmapper.ModelMapper;
+import org.springframework.data.domain.Page;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Service;
+
+import java.util.*;
+import java.util.stream.Collectors;
 
 @Service
 @AllArgsConstructor
@@ -48,10 +44,14 @@ public class QuizServiceImpl implements QuizService {
   @Override
   public Page<QuizDto> findAllQuiz(FindQuizRequest request, PageCriteria pageCriteria) {
     var pageable = PageCriteriaPageableMapper.toPageable(pageCriteria);
-    var response = quizRepository.findAllByConditions(request, pageable).map(quiz -> {
-      var quizDto = modelMapper.map(quiz, QuizDto.class);
-      return quizDto;
-    });
+    var response =
+        quizRepository
+            .findAllByConditions(request, pageable)
+            .map(
+                quiz -> {
+                  var quizDto = modelMapper.map(quiz, QuizDto.class);
+                  return quizDto;
+                });
     return response;
   }
 
@@ -68,8 +68,7 @@ public class QuizServiceImpl implements QuizService {
   @Override
   public QuizDto findById(Long quizId) {
     var quizOpt = quizRepository.findById(quizId);
-    if (quizOpt.isEmpty())
-      return null;
+    if (quizOpt.isEmpty()) return null;
     var quizDto = modelMapper.map(quizOpt.get(), QuizDto.class);
     return quizDto;
   }
@@ -88,9 +87,11 @@ public class QuizServiceImpl implements QuizService {
     quiz.setCreatedDate(new Date());
     quiz.setModifiedDate(new Date());
     quiz.setAuthor(localUser.getUser().getFullName());
-    quiz.getQuestions().forEach(question -> {
-      question.setQuiz(quiz);
-    });
+    quiz.getQuestions()
+        .forEach(
+            question -> {
+              question.setQuiz(quiz);
+            });
     for (Question question : quiz.getQuestions()) {
       question.getAnswers().forEach(answer -> answer.setQuestion(question));
     }
@@ -104,9 +105,11 @@ public class QuizServiceImpl implements QuizService {
     var quiz = modelMapper.map(updateQuizRequest, Quiz.class);
     quiz.setModifiedDate(new Date());
     quiz.setModifiedBy(localUser.getUser().getId());
-    quiz.getQuestions().forEach(question -> {
-      question.setQuiz(quiz);
-    });
+    quiz.getQuestions()
+        .forEach(
+            question -> {
+              question.setQuiz(quiz);
+            });
     for (Question question : quiz.getQuestions()) {
       question.getAnswers().forEach(answer -> answer.setQuestion(question));
     }
@@ -118,18 +121,26 @@ public class QuizServiceImpl implements QuizService {
     LocalUser localUser =
         (LocalUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
     var userId = localUser.getUser().getId();
-    return quizRepository.getQuizTakingHistory(userId).stream().map(item -> {
-      var quizTakingHistoryResponse = modelMapper.map(item, QuizTakingHistoryResponse.class);
-      return quizTakingHistoryResponse;
-    }).collect(Collectors.toList());
+    return quizRepository.getQuizTakingHistory(userId).stream()
+        .map(
+            item -> {
+              var quizTakingHistoryResponse =
+                  modelMapper.map(item, QuizTakingHistoryResponse.class);
+              return quizTakingHistoryResponse;
+            })
+        .collect(Collectors.toList());
   }
 
   @Override
   public QuizResultResponse getQuizResult(ResultQuizCheckingRequest request) {
-    var results = questionRepository.getQuizResult(request.getQuizId()).stream().map(ele -> {
-      var item = modelMapper.map(ele, QuizResultCheckingDto.class);
-      return item;
-    }).collect(Collectors.toList());
+    var results =
+        questionRepository.getQuizResult(request.getQuizId()).stream()
+            .map(
+                ele -> {
+                  var item = modelMapper.map(ele, QuizResultCheckingDto.class);
+                  return item;
+                })
+            .collect(Collectors.toList());
     var response = calculateUserResult(request, results);
     saveToQuizResult(response, request.getQuizId());
     return response;
@@ -146,10 +157,11 @@ public class QuizServiceImpl implements QuizService {
     quizResultRepository.save(quizResult);
   }
 
-  private QuizResultResponse calculateUserResult(ResultQuizCheckingRequest request,
-      List<QuizResultCheckingDto> results) {
-    results.sort(Comparator.comparing(QuizResultCheckingDto::getQuestionId)
-        .thenComparing(QuizResultCheckingDto::getAnswerId));
+  private QuizResultResponse calculateUserResult(
+      ResultQuizCheckingRequest request, List<QuizResultCheckingDto> results) {
+    results.sort(
+        Comparator.comparing(QuizResultCheckingDto::getQuestionId)
+            .thenComparing(QuizResultCheckingDto::getAnswerId));
 
     Map<Long, String> answers = new HashMap();
     int i = 0, numberOfQuestion = 0;
@@ -181,17 +193,13 @@ public class QuizServiceImpl implements QuizService {
       Collections.sort(userSelection.getAnswerIds());
       String answer = "";
       for (Long answerId : userSelection.getAnswerIds()) {
-        if (answer.isEmpty())
-          answer = answerId + "-";
-        else
-          answer = answer + answerId + "-";
+        if (answer.isEmpty()) answer = answerId + "-";
+        else answer = answer + answerId + "-";
       }
       var trueAnswer = answers.get(questionId);
-      if (answer.equals(trueAnswer) == false)
-        numberOfFalse++;
-
+      if (answer.equals(trueAnswer) == false) numberOfFalse++;
     }
-    var score = 100 - (int) (numberOfFalse * 100 / numberOfQuestion);
+    var score = 100 - (numberOfFalse * 100 / numberOfQuestion);
     var response = new QuizResultResponse();
     response.setNumberOfFalse(numberOfFalse);
     response.setNumberOfQuestion(numberOfQuestion);
@@ -205,11 +213,12 @@ public class QuizServiceImpl implements QuizService {
     LocalUser localUser =
         (LocalUser) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
     var userId = localUser.getUser().getId();
-    return quizRepository.findAllByCreatedByAndIsDraft(userId, true).stream().map(quiz -> {
-      var quizOverviewDto = modelMapper.map(quiz, QuizOverviewDto.class);
-      return quizOverviewDto;
-    }).collect(Collectors.toList());
+    return quizRepository.findAllByCreatedByAndIsDraft(userId, true).stream()
+        .map(
+            quiz -> {
+              var quizOverviewDto = modelMapper.map(quiz, QuizOverviewDto.class);
+              return quizOverviewDto;
+            })
+        .collect(Collectors.toList());
   }
-
-
 }
